README.md
A morphological analyzer and generator for 150 Laz nouns. A finite state transducer for 150 Laz (Pazar dialect) nouns, using HFST-dev. It analyzes the nouns fed into the program morphologically, and also generates morphologically correct Laz words when given a root and affix(es). The contributors to this project are Betulay Aras, Aybuke Gungor, and Maya Singhal.